changed code so we can use the RC to change WV_EN parameter

+++ src/modules/rc_update/rc_update.cpp
@@ -291,14 +291,14 @@ void RCUpdate::set_params_from_rc()
 			continue;
 		}
 
-		float rc_val = get_rc_value((rc_channels_s::FUNCTION_PARAM_1 + i), -1.f, 1.f);
+		int rc_val = get_rc_value((rc_channels_s::FUNCTION_PARAM_1 + i), -1.f, 1.f);
 
 		/* Check if the value has changed,
 		 * maybe we need to introduce a more aggressive limit here */
 		if (rc_val > _param_rc_values[i] + FLT_EPSILON || rc_val < _param_rc_values[i] - FLT_EPSILON) {
 			_param_rc_values[i] = rc_val;
-			float param_val = math::constrain(
-						  _rc_parameter_map.value0[i] + _rc_parameter_map.scale[i] * rc_val,
+			int param_val = math::constrain(
+						  int32_t(_rc_parameter_map.value0[i] + _rc_parameter_map.scale[i] * rc_val),
 						  _rc_parameter_map.value_min[i], _rc_parameter_map.value_max[i]);
 
 			param_set(_parameter_handles.rc_param[i], &param_val);
